requirements.txt
customtkinter==5.2.1
pyyaml==6.0
pyinstaller==6.1.0
